python_console_app/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
drf-yasg==1.21.10
flake8==7.2.0
inflection==0.5.1
iniconfig==2.1.0
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
pycodestyle==2.13.0
pyflakes==3.3.2
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
django-cors-headers==4.7.0
# Legacy deps retained but no longer used by default flow:
youtube-transcript-api==0.6.2
pytube==15.0.0
yt-dlp==2025.1.26
# New dependencies for PDF reading and image creation
PyPDF2==3.0.1
Pillow==10.4.0
